fix: geçmiş pozisyonda kesirli lot desteği (lots Integer→BigDecimal + V31 migrasyon, 400 hatası giderildi), tutar modu lot sayısına göre dönüşsün

+++ backend/src/main/java/com/finansportali/backend/dto/request/HistoricalPositionRequest.java
@@ -23,7 +23,7 @@ public record HistoricalPositionRequest(
 
     @NotNull(message = "Lots is required")
     @Positive(message = "Lots must be positive")
-    Integer lots,
+    BigDecimal lots,
 
     @NotBlank(message = "Currency is required")
     String currency
